Moved BookingFinder to separate module
