+++ finder_template.csv
@@ -1,0 +1,4 @@
+3TE, 3TW, 7T, 3TM, Peter Hobden, Allison Cooper
+1, 0, 0, 0, 0, 1
+1, 0, 0, 0, 0, 1
+0, 0, 1, 0, 0, 1
